# src/docs/swagger.yaml
openapi: 3.0.0
info:
    title: Scope manager API
    version: 1.0.0
    description: |
        REST API for managing organizations, their roles, element fields
        and agreement fields within Scope manager API.

        Each organization is the main container in the system: it groups elements,
        defines the vocabulary they must follow and establishes the available roles.
servers:
    - url: /
...
paths:
...
    /api/v1/organizations/{orgName}:
        parameters:
            - $ref: '#/components/parameters/orgName'
        get:
            tags:
                - Organizations
            summary: Get an organization by name
            operationId: getOrganizationByName
            security:
                - BearerAuth: []
            description: |
                Returns the full organization identified by its unique name,
                including its roles, element fields, agreement fields and members count.
                Requires an authenticated user that is either a member of the organization
                or has system ADMIN role.
            responses:
                '200':
                    description: Organization found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                            example:
                                success: true
                                message: OK
                                httpStatus: 200
                                appCode: SUCCESS
                                data:
                                    _id: 665a1b2c3d4e5f6a7b8c9d0e
                                    name: universidad-sevilla
                                    displayName: Universidad de Sevilla
                                    description: Organization for Universidad de Sevilla
                                    createdBy: 665d4e5f6a7b8c9d0e1f2a3b
                                    elementFields: []
                                    agreementFields: []
                                    roles:
                                        - _id: 665b2c3d4e5f6a7b8c9d0e1f
                                          name: admin
                                          description: Organization Administrator
                                    members: 1
                                    createdAt: '2025-06-01T10:00:00.000Z'
                                    updatedAt: '2025-06-01T10:00:00.000Z'
                                    __v: 0
                                error: null
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/OrganizationNotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        put:
            tags:
                - Organizations
            summary: Update an organization
            operationId: updateOrganization
            security:
                - BearerAuth: []
            description: |
                Updates the name, displayName and description of an existing organization.
                Requires organization admin role or system ADMIN role.
                If the name is changed, the new name must be unique in the system.
                The request body is validated with the same required fields as creation.
            requestBody:
                $ref: '#/components/requestBodies/OrganizationBody'
            responses:
                '200':
                    description: Organization updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrganizationResponse'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '409':
                    $ref: '#/components/responses/OrganizationConflict'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            tags:
                - Organizations
            summary: Delete an organization
            operationId: deleteOrganization
            security:
                - BearerAuth: []
            description: |
                Permanently deletes the organization and all its associated data.
                Requires organization admin role or system ADMIN role.
                Memberships and elements belonging to the organization are removed before the organization.
            responses:
                '200':
                    description: Organization deleted successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NullResponse'
                            example:
                                success: true
                                message: Organization deleted
                                httpStatus: 200
                                appCode: SUCCESS
                                data: null
                                error: null
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ──────────────────────────────────────────────
    # Roles
    # ──────────────────────────────────────────────
...
components:
...
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: JWT bearer token

    # ── Parameters ────────────────────────────────
    parameters:
        orgName:
            name: orgName
            in: path
            required: true
            description: Unique name of the organization
            schema:
                type: string
            example: universidad-sevilla
...
    requestBodies:
        OrganizationBody:
            description: |
                Data to create or update an organization.
                `name` is required, must be 3-100 characters and can only contain letters, numbers and hyphens.
                `displayName` is optional and must be at most 200 characters.
                `description` is required and must be 3-500 characters.
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/OrganizationInput'
                    example:
                        name: universidad-sevilla
                        displayName: Universidad de Sevilla
                        description: Organization for Universidad de Sevilla
...
    responses:
        ValidationError:
            description: Validation error in the submitted data
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Validation error
                        httpStatus: 400
                        appCode: VALIDATION_ERROR
                        data: null
                        error:
                            message: Validation error
                            details: null
        NotFound:
            description: The requested resource does not exist
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Resource not found
                        httpStatus: 404
                        appCode: NOT_FOUND
                        data: null
                        error:
                            message: Resource not found
                            details: null
...
        OrganizationNotFound:
            description: The organization does not exist
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: "Organization with name 'universidad-sevilla' not found"
                        httpStatus: 404
                        appCode: NOT_FOUND
                        data: null
                        error:
                            message: "Organization with name 'universidad-sevilla' not found"
                            details: null
        OrganizationConflict:
            description: An organization with that name already exists
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: An organization with that name already exists
                        httpStatus: 409
                        appCode: DUPLICATE_KEY
                        data: null
                        error:
                            message: An organization with that name already exists
                            details: null
...
        Unauthorized:
            description: Missing or invalid authentication token
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Authorization header missing or malformed
                        httpStatus: 401
                        appCode: UNAUTHORIZED
                        data: null
                        error:
                            message: Authorization header missing or malformed
                            details: null
        Forbidden:
            description: Insufficient permissions to perform this action
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: You do not have permission to perform that action
                        httpStatus: 403
                        appCode: FORBIDDEN
                        data: null
                        error:
                            message: You do not have permission to perform that action
                            details: null
        InternalError:
            description: Internal server error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Internal server error
                        httpStatus: 500
                        appCode: UNKNOWN_ERROR
                        data: null
                        error:
                            message: Internal server error
                            details: null

    # ── Schemas ───────────────────────────────────
    schemas:
...
        SuccessResponse:
            type: object
            description: Standard wrapper for all successful API responses
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: OK
                httpStatus:
                    type: integer
                    example: 200
                appCode:
                    type: string
                    example: SUCCESS
                error:
                    type: object
                    nullable: true
                    example: null
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error
        OrganizationResponse:
            allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - type: object
                  properties:
                      data:
                          $ref: '#/components/schemas/Organization'
...
        NullResponse:
            allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - type: object
                  properties:
                      data:
                          type: object
                          nullable: true
                          example: null
...
        ErrorResponse:
            type: object
            description: Standard wrapper for all error API responses
            properties:
                success:
                    type: boolean
                    example: false
                message:
                    type: string
                httpStatus:
                    type: integer
                appCode:
                    type: string
                data:
                    type: object
                    nullable: true
                    example: null
                error:
                    type: object
                    properties:
                        message:
                            type: string
                        details:
                            description: Additional error information (may be null)
                            type: object
                            nullable: true
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error
...
        Organization:
            type: object
            description: |
                Main container of the system. Defines the structure of its elements,
                the structure of its agreements and the available roles.
            required:
                - _id
                - name
                - description
                - createdBy
                - elementFields
                - agreementFields
                - roles
            properties:
                _id:
                    type: string
                    description: MongoDB internal identifier (ObjectId)
                    example: 665a1b2c3d4e5f6a7b8c9d0e
                name:
                    type: string
                    description: Unique organization name (used as identifier in the API). Only letters, numbers and hyphens allowed.
                    pattern: '^[a-zA-Z0-9-]+$'
                    minLength: 3
                    maxLength: 100
                    example: universidad-sevilla
                displayName:
                    type: string
                    description: Human-readable display name. Can contain any characters. Defaults to empty string.
                    maxLength: 200
                    default: ''
                    example: Universidad de Sevilla
                description:
                    type: string
                    description: Organization description
                    minLength: 3
                    maxLength: 500
                    example: Organization for Universidad de Sevilla
                createdBy:
                    type: string
                    description: ObjectId of the user that created the organization
                    example: 665d4e5f6a7b8c9d0e1f2a3b
                elementFields:
                    type: array
                    description: Fields that define the structure of elements
                    items:
                        $ref: '#/components/schemas/ElementField'
                agreementFields:
                    type: array
                    description: Fields that define the structure of agreements
                    items:
                        $ref: '#/components/schemas/AgreementField'
                roles:
                    type: array
                    description: Available roles within the organization
                    items:
                        $ref: '#/components/schemas/Role'
                members:
                    type: integer
                    description: Number of members in the organization
                    example: 1
                createdAt:
                    type: string
                    format: date-time
                    description: Creation date (auto-generated by Mongoose)
                updatedAt:
                    type: string
                    format: date-time
                    description: Last update date (auto-generated by Mongoose)
                __v:
                    type: integer
                    description: Mongoose version key
                    example: 0
        OrganizationInput:
            type: object
            description: Data required to create or update an organization
            required:
                - name
                - description
            properties:
                name:
                    type: string
                    description: Unique organization name. Only letters, numbers and hyphens allowed.
                    pattern: '^[a-zA-Z0-9-]+$'
                    minLength: 3
                    maxLength: 100
                    example: universidad-sevilla
                displayName:
                    type: string
                    description: Human-readable display name. Can contain any characters.
                    maxLength: 200
                    example: Universidad de Sevilla
                description:
                    type: string
                    description: Organization description
                    minLength: 3
                    maxLength: 500
                    example: Organization for Universidad de Sevilla
...
        Role:
            type: object
            description: Role within an organization (e.g. admin, teacher, student)
            required:
                - _id
                - name
                - description
            properties:
                _id:
                    type: string
                    description: MongoDB internal identifier (subdocument ObjectId)
                    example: 665b2c3d4e5f6a7b8c9d0e1f
                name:
                    type: string
                    description: Role name (unique within the organization)
                    minLength: 3
                    maxLength: 100
                    example: admin
                description:
                    type: string
                    description: Role description
                    minLength: 3
                    maxLength: 500
                    example: Organization administrator
...
        FieldBase:
            type: object
            required:
                - name
                - description
                - type
            properties:
                name:
                    type: string
                    description: Field name (unique within its array)
                    minLength: 3
                    maxLength: 100
                description:
                    type: string
                    description: Field description
                    minLength: 3
                    maxLength: 500
                type:
                    type: string
                    description: Field type
                    enum:
                        - string
                        - enum
                        - number

        # ── ElementField ─────────────────────────
        ElementField:
            description: |
                Field that defines part of the structure of an organization's elements.
                If `type` is `enum`, `value` contains the array of allowed values.
                For any other type, `value` is not present.
            oneOf:
                - title: Enum element field
                  allOf:
                      - $ref: '#/components/schemas/FieldBase'
                      - type: object
                        required:
                            - value
                        properties:
                            value:
                                type: array
                                description: Allowed values for the enum field
                                items:
                                    type: string
                                example:
                                    - PSG1
                                    - PSG2
                                    - PSG3
                  example:
                      name: class
                      description: Project class
                      type: enum
                      value:
                          - PSG1
                          - PSG2
                          - PSG3
                - title: Non-enum element field
                  allOf:
                      - $ref: '#/components/schemas/FieldBase'
                  example:
                      name: title
                      description: Element title
                      type: string
...
        AgreementField:
            description: |
                Field that defines part of the structure of an organization's agreements.
                If `type` is `enum`, `value` contains the array of allowed values.
                For any other type, `value` is not present.
            oneOf:
                - title: Enum agreement field
                  allOf:
                      - $ref: '#/components/schemas/FieldBase'
                      - type: object
                        required:
                            - value
                        properties:
                            value:
                                type: array
                                description: Allowed values for the enum field
                                items:
                                    type: string
                                example:
                                    - gold
                                    - silver
                                    - bronze
                  example:
                      name: guaranteeLevel
                      description: Service level guarantee tier
                      type: enum
                      value:
                          - gold
                          - silver
                          - bronze
                - title: Non-enum agreement field
                  allOf:
                      - $ref: '#/components/schemas/FieldBase'
                  example:
                      name: minResponseTime
                      description: Minimum response time in hours
                      type: number
...